working full, except foodlist md and xs viewport.

+++ app/Http/Services/NewsService.php
@@ -7,28 +7,25 @@ use App\Models\News;
 use App\Models\NewsLocalization;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Storage;
 
 class NewsService
 {
     public function insert(Request $request) {
-        $postData = $request->all('news')['news'];
+        $postData = $request->all();
+        unset($postData['_token']);
 //        dd($postData);
 //        dump($postData);
         $imageService = new ImageService();
 
         if(!isset($postData['has_image'])) {
             unset($postData['image']);
-            $file       = $request->file('news.image');
-            $storedFile = $imageService->put($file);
-
-            $postData['image_id']     = $storedFile->id;
-            $postData['img_filename'] = $storedFile->filename;
-            $postData['img_path']     = $storedFile->path;
-            $imageId                  = $storedFile->id;
-        } else {
-
-            $imageId = $postData['image_id'];
+            $file                        = $request->file('image');
+            $storedFile                  = $imageService->put($file);
+            $postData['image_file_name'] = $storedFile->image_file_name;
+            $postData['image_path']      = $storedFile->image_path;
         }
+//            dd($postData);
 
         if(!isset($postData['is_active'])) {
             $postData['is_active'] = 0;
@@ -36,44 +33,9 @@ class NewsService
 
         $news = News::updateOrCreate(
             ['id' => $postData['id']],
-            [
-                'category_id' => $postData['category_id'],
-                'start'       => $postData['start'],
-                'end'         => $postData['end'],
-                'image_id'    => $imageId,
-            ],
+            $postData,
         );
 //        dump('$news', $news);
-
-        foreach($postData['labels'] as $key => $contentType) {
-//            dump('$contentType', $contentType);
-//            dump('$key', $key);
-            foreach($contentType as $localization) {
-//                dump('$localization', $localization);
-                $localizationUpdated = Localization::updateOrCreate(
-                    ['id' => $localization['localization_id']],
-                    [
-                        'label'           => $localization['label'],
-                        'language_id'     => $localization['language_id'],
-                        'content_type_id' => $localization['content_type_id'],
-                    ]);
-                if($localization['localization_id'] == null) {
-                    $localization['localization_id'] = $localizationUpdated->id;
-                }
-//            dump($localizationUpdated);
-//            dump($news->id);
-
-                $newsLocalization = NewsLocalization::updateOrCreate(
-                    ['id' => $localization['news_localization_id']],
-                    [
-                        'news_id'         => $news->id,
-                        'localization_id' => $localizationUpdated->id,
-                    ]);
-//            dump($newsLocalization);
-
-            }
-        }
-
 //        dd('HIIIIERR');
 
 
@@ -120,47 +82,17 @@ class NewsService
         return $news;
     }
 
-    public function allGermanHeading() {
-//        dump('NewsService->allGermanHeading()');
+    public function all() {
+        return News::get();
+    }
 
-        $newsArray = [];
-        $postNews  = [];
-        $newsRaw   = News::has('newsLocalizations.localizationGerman')
-            ->with([
-                'category.categoryLocalizationGerman.localizationGerman',
-                'newsLocalizationsGerman.localizationGerman.language',
-                'newsLocalizationsGerman.localizationGerman.contentType'
-            ])->get();
-//        dump('$newsRaw',$newsRaw);
-        foreach($newsRaw as $news) {
-//            dd($news);
+    public function allActive() {
+        $nowToDate = now()->toDateString();
+        return News::where([
+            ['display_from','>=', $nowToDate],
+            ['display_to','>=', $nowToDate],
+        ])->get();
 
-            $newsArray['id']    = $news->id;
-            $newsArray['start'] = $news->start;
-            $newsArray['end']   = $news->end;
-            foreach($news->newsLocalizationsGerman as $item) {
-//                dump('$item',$item);
-//                dump('$news',$news);
-
-//                dd( $news->category->categoryLocalizationGerman);
-
-                $newsArray['localizations'][$item->localizationGerman->contentType->label] = [
-                    'id'              => $item->id,
-                    'label'           => $item->localizationGerman->label,
-                    'category_label'  => $news->category->categoryLocalizationGerman->localizationGerman->label,
-                    'iso_code'        => $item->localizationGerman->language->iso_code,
-                    'category_id'     => $news->category_id,
-                    'localization_id' => $item->localizationGerman->id,
-                    'content_type_id' => $item->localizationGerman->content_type_id,
-                    'language_id'     => $item->localizationGerman->language_id,
-                ];
-
-            }
-            $postNews[] = $newsArray;
-        }
-//        dump($postNews);
-//        dd('HIIERR');
-        return $this->toObject($postNews);
     }
 
     public function allDependingOnLanguage() {
@@ -240,40 +172,8 @@ class NewsService
         return $this->toObject($postnews);
     }
 
-    public function single($id) {
-        $newsRaw = News::with([
-            'newsLocalizations.localization.language',
-            'newsLocalizations.localization.contentType',
-            'image'
-        ])->whereId($id)->first();
-
-//        dump($newsRaw->newsLocalizations);
-
-        $news['id']    = $newsRaw->id;
-        $news['start'] = $newsRaw->start ?? null;
-        $news['end']   = $newsRaw->end ?? null;
-        if($newsRaw->image != null) {
-
-            $news['image_id']   = $newsRaw->image->id ?? null;
-            $news['image_id']   = $newsRaw->image->id ?? null;
-            $news['image_path'] = $newsRaw->image->path ?? null;
-        }
-
-        foreach($newsRaw->newsLocalizations as $item) {
-//            dump($item->localization->contentType->label);
-            $news['localizations'][$item->localization->language->label][$item->localization->contentType->label] = [
-                'id'              => $item->id,
-                'label'           => $item->localization->label,
-                'iso_code'        => $item->localization->language->iso_code,
-                'category_id'     => $newsRaw->category_id,
-                'localization_id' => $item->localization->id,
-                'content_type_id' => $item->localization->content_type_id,
-                'language_id'     => $item->localization->language_id,
-            ];
-        }
-//        dump($news);
-//        dd('HIIIERRR');
-        return $news;
+    public function single($id = null) {
+        return News::whereId($id)->first();
     }
 
     public function singleRaw($id = null) {
@@ -283,16 +183,24 @@ class NewsService
         ])->whereId($id)->first();
     }
 
-    public function delete($id) {
-        $news = $this->singleRaw($id);
-        foreach($news->newsLocalizations as $newsLocalization) {
-            Localization::whereId($newsLocalization->localization->id)->delete();
-            NewsLocalization::whereId($newsLocalization->id)->delete();
+    public function delete($id = null) {
+        if($id) {
+            $news = $this->single($id);
+            if($news) {
+                $this->deleteImage($id);
+                $news->delete();
+            }
         }
-        $news->delete();
+
     }
 
-    public function toObject($array) {
-        return json_decode(json_encode($array));
+    public function deleteImage($id = null) {
+        $news                  = $this->single($id);
+        $news->image_file_name = null;
+        $news->image_path      = null;
+        $news->save();
+        if(Storage::exists('public/images/upload/news/' . $news->image_file_name)) {
+            Storage::delete('public/images/upload/news/' . $news->image_file_name);
+        }
     }
 }
